Refactor application structure by removing unused components and styles. Reintroduce Navbar with updated styles and functionality. Enhance routing and page components for improved user experience.
