+++ app/src/App.js
@@ -1,44 +1,37 @@
-import './App.css';
-import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+import "./App.css";
+import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 
+import Navbar from "./components/NavBar";
+import Home from "./pages/Home";
 
-import Navbar from './components/Navbar';
-import Home from './pages/Home';
+import Will from "./pages/Will/WillChat";
+import Ben from "./pages/Ben/BenChat";
 
-import Will from './pages/Will/Will-chat';
-import Ben from './pages/BenChat';
+import BenImages from "./pages/Ben/BenImages";
+import WillImages from "./pages/Will/WillImages";
 
-import BenImages from './pages/BenImages';
-import WillImages from './pages/WillImages';
+import BenAPI from "./pages/Ben/BenAPI";
+import WillApi from "./pages/Will/WillApi";
 
-
-import BenAPI from './pages/BenAPI';
-import WillApi from './pages/Will/WillApi';
-
-function App() {  
-
-
+function App() {
   return (
     <div className="App">
-    <header className="App-header">
-    <Router>
-      <Navbar />
-        <Routes>
-          <Route path="/Home" element={<Home />} />
-          
-          <Route path="/will" element={<Will />} />
-          <Route path="/ben" element={<Ben />} />
-          <Route path="/willapi" element={<WillApi />} />
-          <Route path="/BenAPI" element={<BenAPI />} />
-          <Route path="/WillImages" element={<WillImages />} />
-          <Route path="/BenImages" element={<BenImages />} />
-        </Routes>
-        
-    </Router>
-        
-  
-    </header>
-      </div>
+      <header className="App-header">
+        <Router>
+          <Navbar />
+          <Routes>
+            <Route path="/Home" element={<Home />} />
+
+            <Route path="/will" element={<Will />} />
+            <Route path="/ben" element={<Ben />} />
+            <Route path="/willapi" element={<WillApi />} />
+            <Route path="/BenAPI" element={<BenAPI />} />
+            <Route path="/WillImages" element={<WillImages />} />
+            <Route path="/BenImages" element={<BenImages />} />
+          </Routes>
+        </Router>
+      </header>
+    </div>
   );
 }
 export default App;
